+++ ivy/src/main/scala/sbt/mavenint/PomExtraDependencyAttributes.scala
@@ -1,4 +1,4 @@
-package org.apache.maven.repository.internal
+package sbt.mavenint
 
 import java.util.Properties
 import java.util.regex.Pattern
@@ -31,7 +31,7 @@ object PomExtraDependencyAttributes {
    *         A map of module id to extra dependency attributes associated with dependencies on that module.
    */
   def readFromAether(props: java.util.Map[String, AnyRef]): Map[ModuleRevisionId, Map[String, String]] = {
-    import collection.JavaConverters._
+    import scala.collection.JavaConverters._
     (props.asScala get ExtraAttributesKey) match {
       case None => Map.empty
       case Some(str) =>
@@ -71,7 +71,7 @@ object PomExtraDependencyAttributes {
     }
 
   def qualifiedExtra(item: ExtendableItem): Map[String, String] = {
-    import collection.JavaConverters._
+    import scala.collection.JavaConverters._
     item.getQualifiedExtraAttributes.asInstanceOf[java.util.Map[String, String]].asScala.toMap
   }
   def filterCustomExtra(item: ExtendableItem, include: Boolean): Map[String, String] =
@@ -83,7 +83,7 @@ object PomExtraDependencyAttributes {
   // This makes the id suitable as a key to associate a dependency parsed from a <dependency> element
   //  with the extra attributes from the <properties> section
   def simplify(id: ModuleRevisionId): ModuleRevisionId = {
-    import collection.JavaConverters._
+    import scala.collection.JavaConverters._
     ModuleRevisionId.newInstance(id.getOrganisation, id.getName, id.getBranch, id.getRevision, filterCustomExtra(id, include = false).asJava)
   }
