+++ sbt-maven-resolver/src/main/scala/sbt/mavenint/MavenResolutionException.scala
@@ -1,0 +1,4 @@
+package sbt.mavenint
+
+/** An exception we can throw if we encounter issues. */
+class MavenResolutionException(msg: String) extends RuntimeException(msg) {}
